Explicitly disable server name indication
Server name indication is a way for a server to present
multiple TLS certificates from a single IP address. As of OTP
20 the SSL application passed back the SNI information to the
client so that it can be validated.

We restore the previous behavior by disabling this
functionality explicitly.

The SNI disable option has existed since at least
OTP 17 so it ought to be safe for several previous
OTP releases.